+++ client/app/review/ReviewContainer.jsx
@@ -7,7 +7,7 @@ import * as reviewActions from './reviewActions';
 import * as itemActions from '../items/itemActions';
 
 import Header from '../../components/Header';
-import PageTemplate from '../../components/PageTemplate';
+import ProgressBar from './ReviewProgressBar';
 
 export const REVIEW_TYPE = {
   EASY: 'easy',
@@ -26,21 +26,6 @@ const SessionPage = ({ children, title }) => (
       </div>
     </div>
   </Header>
-);
-
-const ProgressBar = ({ progress }) => (
-  <div className="progress">
-    <div
-      className="progress-bar"
-      role="progressbar"
-      aria-valuenow={progress}
-      aria-valuemin="0"
-      aria-valuemax="100"
-      style={{ width: `${progress}%`, backgroundColor: '#1574fb' }}
-    >
-      <span className="sr-only">{progress}% Complete</span>
-    </div>
-  </div>
 );
 
 const SessionResultItem = ({ item }) => (
